Introduced better resolution

+++ sonoff/xsns_04_snfsc.ino
@@ -55,7 +55,7 @@
 
 #define XSNS_04             4
 
-uint16_t sc_value[5] = { 0 };
+float sc_value[5] = { 0 };
 
 void SonoffScSend(const char *data)
 {
@@ -75,22 +75,22 @@ void SonoffScSerialInput(char *rcvstat)
 {
   char *p;
   char *str;
-  uint16_t value[5] = { 0 };
+  float value[5] = { 0 };
 
   AddLog_P2(LOG_LEVEL_DEBUG, PSTR(D_LOG_SERIAL D_RECEIVED " %s"), rcvstat);
 
   if (!strncasecmp_P(rcvstat, PSTR("AT+UPDATE="), 10)) {
     int8_t i = -1;
     for (str = strtok_r(rcvstat, ":", &p); str && i < 5; str = strtok_r(nullptr, ":", &p)) {
-      value[i++] = atoi(str);
+      value[i++] = atof(str);
     }
     if (value[0] > 0) {
       for (uint32_t i = 0; i < 5; i++) {
         sc_value[i] = value[i];
       }
-      sc_value[2] = (11 - sc_value[2]) * 10;  // Invert light level
+      sc_value[2] = (uint16_t)(((uint16_t)11 - (uint16_t)sc_value[2]) * 10);  // Invert light level
       sc_value[3] *= 10;
-      sc_value[4] = (11 - sc_value[4]) * 10;  // Invert dust level
+      sc_value[4] = (uint16_t)(((uint16_t)11 - (uint16_t)sc_value[4]) * 10);  // Invert dust level
       SonoffScSend("AT+SEND=ok");
     } else {
       SonoffScSend("AT+SEND=fail");
@@ -121,7 +121,7 @@ void SonoffScShow(bool json)
 
     if (json) {
       ResponseAppend_P(PSTR(",\"SonoffSC\":{\"" D_JSON_TEMPERATURE "\":%s,\"" D_JSON_HUMIDITY "\":%s,\"" D_JSON_LIGHT "\":%d,\"" D_JSON_NOISE "\":%d,\"" D_JSON_AIRQUALITY "\":%d}"),
-        temperature, humidity, sc_value[2], sc_value[3], sc_value[4]);
+        temperature, humidity, (uint16_t)sc_value[2], (uint16_t)sc_value[3], (uint16_t)sc_value[4]);
 #ifdef USE_DOMOTICZ
       if (0 == tele_period) {
         DomoticzTempHumSensor(temperature, humidity);
@@ -142,7 +142,7 @@ void SonoffScShow(bool json)
     } else {
       WSContentSend_PD(HTTP_SNS_TEMP, "", temperature, TempUnit());
       WSContentSend_PD(HTTP_SNS_HUM, "", humidity);
-      WSContentSend_PD(HTTP_SNS_SCPLUS, sc_value[2], sc_value[3], sc_value[4]);
+      WSContentSend_PD(HTTP_SNS_SCPLUS, (uint16_t)sc_value[2], (uint16_t)sc_value[3], (uint16_t)sc_value[4]);
 #endif  // USE_WEBSERVER
     }
   }
